+++ CHANGES.md
@@ -5,7 +5,7 @@ This file records the changes in each m3api release.
 The annotated tag (and GitHub release) for each version also lists the changes,
 but this file may sometimes contain later improvements (e.g. typo fixes).
 
-## next (not yet released)
+## v0.7.3 (2023-04-11)
 
 - The new `errorHandlers` option can be used to customize error handling.
   It is only part of the internal interface, not of the stable, public interface.